Fixes to make the tess run successfully on Travis.
The main issue was of installing yaml dependency in blender's python, as
that does not ship with headers and yaml installation would fail with
`Python.h` not found error.

The solution is to:
1. Install yaml in the Travis CI venv.
2. Copy the yaml from Travis CI venv site packages to blender's modules
directory, where blender automatically looks for python depdencies! (Did
not know this trick until today)
3. The other change required was to use PYTHONPATH=. when running
tests.py as the cavity tutorial test imports the base class
`TestFoamTutorial` which is part of tests package.